Move member tests and add one for maps
